Cleanup Docker contianers, images
This patch cleanup the leftover docker containers, images
and also add NEXUS_URL to the script to publish fabric
docker images

+++ jjb/fabric/shell/include-raw-FIT-JAVA-e2e.sh
@@ -1,7 +1,9 @@
 #!/bin/bash -e
 set -o pipefail
 
-if [ "$GERRIT_BRANCH" != "release-1.0" ] || [ "$ARCH" != "s390x" ] || [ "$ARCH" != "ppc64le" ]; then
+echo "---------> GERRIT_BRANCH" $GERRIT_BRANCH
+echo "---------> ARCH" $ARCH
+if [ "$GERRIT_BRANCH" != "release-1.0" ] && [ "$ARCH" != "s390x" ] && [ "$ARCH" != "ppc64le" ]; then
       echo "----> $GERRIT_BRANCH"
       # Move to fabric-sdk-java repository and execute end-to-end tests
       rm -rf ${WORKSPACE}/gopath/src/github.com/hyperledger/fabric-sdk-java
